Create Shiftbrite_Tuner_with_LCD_v0dot6.ino
This was how I learned to use Serial.print to write to an LCD screen, and I combined it with some of my arduino code from the Sheri_sign project. I used the code to pick a few colors from the array (and shift them onto a display board, see pictures in the link on the README), so that I could then use potentiometers to match the hue and shade on another Shiftbrite, in order to improve my color array, which should improve other projects.
Initially I would just keep shifting colors onto the display boards, and note the values I wanted on paper, and then compile all the new values into the color array for a new version of the tuner code.
Now I've got nice definition on the array, so I'll use it to add even more hues to the array!